requirements.txt
langchain-openai 
PyPDF2 
python-dotenv 
langchain 
langchain-community 
langchain-anthropic 
langchain-google-genai 
langchain-google-firestore 
firestore   
tiktoken  
sentence-transformers 
bs4  
firecrawl-py 
langchainhub 
wikipedia 
tavily-python 
langchain_mistralai
langchain_huggingface
Ipython
